Kyle's side of the completed first draft.
I saw that the State array was starting on 0, so I changed a few lines
on Calculator.
Tested and fixed State and its sorting methods, as well as compareCFR's
errors.

What we have here will get by, but there are more efficient ways of
sorting the arrays which require the renovation of SinglyLinkedList. I
hope to make such changes afterwards, but I don't want to change the UML
diagram too much before our soft deadline.

I am hoping to reassess the UML diagram, come up with a way to sort the
things within the SinglyLinkedList, and remove the test classes.